+++ src/molecules/ElectricityChart/ElectricityChart.js
@@ -10,10 +10,11 @@ import { DateTime } from 'luxon';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
 
-const GAS_EXCHANGES_CHART = gql`
-	query gasExchanges($resolution: TimeSpan $timePeriod: TimePeriodInput!) {
-		gasExchanges(resolution: $resolution, timePeriod: $timePeriod) {
-			received
+const ELECTRIC_RECEIVED_CHART = gql`
+	query electricityExchangesChart($resolution: TimeSpan $timePeriod: TimePeriodInput!) {
+		electricityExchanges(resolution: $resolution, timePeriod: $timePeriod) {
+			received(unit:WATT_HOUR)
+			delivered(unit:WATT_HOUR)
 			dataPointsCount
 			period {
 				start
@@ -25,12 +26,12 @@ const GAS_EXCHANGES_CHART = gql`
 
 const now = DateTime.now();
 
-export default function GasReceivedChart() {
-	const { loading, error, data } = useQuery(GAS_EXCHANGES_CHART, {
+export default function ElectricityReceivedChart({ resolution }) {
+	const { loading, error, data } = useQuery(ELECTRIC_RECEIVED_CHART, {
 		variables: {
-			resolution: 'TWO_HOURS',
+			resolution: resolution || 'FIVE_MINUTES',
 			timePeriod: {
-				start: Math.floor(now.minus({ days: 4 }).toSeconds()),
+				start: Math.floor(now.minus({ hours: 8 }).toSeconds()),
 				end: Math.floor(now.toSeconds())
 			}
 		}
@@ -40,7 +41,7 @@ export default function GasReceivedChart() {
 	if (error) return <Message type="error">{error.message}</Message>;
 
 	return (
-		<div className="gas-usage-container">
+		<div>
 			<HighchartsReact
 				highcharts={Highcharts}
 				options={{
@@ -55,15 +56,15 @@ export default function GasReceivedChart() {
 						type: 'datetime',
 						lineColor: 'hsla(var(--color-secondary-shade-3-h), var(--color-secondary-shade-3-s), var(--color-secondary-shade-3-l), .4)',
 						tickColor: 'hsla(var(--color-secondary-shade-3-h), var(--color-secondary-shade-3-s), var(--color-secondary-shade-3-l), .4)',
-						plotBands: data.gasExchanges.filter(exchange => exchange.dataPointsCount === 0).map(exchange => ({
+						plotBands: data.electricityExchanges.filter(exchange => exchange.dataPointsCount === 0).map(exchange => ({
 							color: 'hsla(var(--color-secondary-shade-1-h), var(--color-secondary-shade-1-s), var(--color-secondary-shade-1-l), 0.1)', // Color value
-							from: exchange.period.start * 1000,
-							to: exchange.period.end * 1000
+							from: exchange.period.start * 1000, // Start of the plot band
+							to: exchange.period.end * 1000 // End of the plot band
 						})),
 					},
 					yAxis: {
 						title: {
-							text: 'm³'
+							text: 'Wh'
 						},
 						gridLineColor: 'var(--color-secondary-shade-2)',
 					},
@@ -71,11 +72,16 @@ export default function GasReceivedChart() {
 						useUTC: false
 					},
 					series: [{
-						name: 'm³',
-						type: 'column',
+						name: 'Wh',
 						showInLegend: false,
-						data: data.gasExchanges.map(usage => [usage.period.end * 1000, usage.received]),
-						color: 'hsla(var(--color-gas-h), var(--color-gas-s), var(--color-gas-l), .6)'
+						data: data.electricityExchanges.map(usage => [usage.period.end * 1000, usage.received]),
+						color: 'hsla(var(--color-electricity-received-h), var(--color-electricity-received-s), var(--color-electricity-received-l), .6)'
+					},
+					{
+						name: 'Wh',
+						showInLegend: false,
+						data: data.electricityExchanges.map(usage => [usage.period.end * 1000, usage.delivered]),
+						color: 'hsla(var(--color-electricity-delivered-h), var(--color-electricity-delivered-s), var(--color-electricity-delivered-l), .6)'
 					}]
 				}}
 			/>